Move Firmware class to IKEA.
